emu86: some more refactoring

+++ dis86/src/emu86/interrupt.rs
@@ -1,7 +1,7 @@
 use super::machine::*;
 
 impl Machine {
-  pub fn interrupt(&mut self, num: u16) {
+  pub fn interrupt(&mut self, num: u8) {
     match num {
       0x21 => self.interrupt_0x21(),
       _ => panic!("unimplmented interrupt"),
